[pytest] Secede `TestbedInfo` from conftest.py
Let's move class `TestbedInfo` out of the topmost `conftest.py` and save
it under `tests/common/testbed.py`:
1. Since it is a common infrastructure functionality, secede it from
`conftest.py` could make it more accessible and make `conftest.py`
simpler.
2. Make it a runnable script as a tool could do the testbed file
conversion from `CSV` to `YAML`.
